chore: refactoring the app_routing widget (#778)
