+++ lib/app/view/app_routing.dart
@@ -127,7 +127,7 @@ class _MaterialApp extends StatelessWidget {
             name: 'scout',
             path: '/scout',
             builder: (BuildContext context, GoRouterState state) {
-              Global().page = 'scout';
+              Global().pageName = 'scout';
               return BlocProvider(
                 create: (BuildContext context) => ScoutPageBloc(
                   tokenRepository: context.read<TokensRepository>(),
@@ -151,7 +151,7 @@ class _MaterialApp extends StatelessWidget {
                 name: 'athlete',
                 path: 'athlete/:id',
                 builder: (BuildContext context, GoRouterState state) {
-                  Global().page = 'athlete';
+                  Global().pageName = 'athlete';
                   return AthletePage(athlete: _toAthlete(state.params['id']!));
                 },
               ),
@@ -161,7 +161,7 @@ class _MaterialApp extends StatelessWidget {
             name: 'trade',
             path: '/trade',
             builder: (BuildContext context, GoRouterState state) {
-              Global().page = 'trade';
+              Global().pageName = 'trade';
               return BlocProvider(
                 create: (BuildContext context) => TradePageBloc(
                   walletRepository: context.read<WalletRepository>(),
@@ -179,7 +179,7 @@ class _MaterialApp extends StatelessWidget {
             name: 'pool',
             path: '/pool',
             builder: (BuildContext context, GoRouterState state) {
-              Global().page = 'pool';
+              Global().pageName = 'pool';
               return BlocProvider(
                 create: (BuildContext context) => AddLiquidityBloc(
                   walletRepository: context.read<WalletRepository>(),
@@ -201,7 +201,7 @@ class _MaterialApp extends StatelessWidget {
             name: 'farm',
             path: '/farm',
             builder: (BuildContext context, GoRouterState state) {
-              Global().page = 'farm';
+              Global().pageName = 'farm';
               return BlocProvider(
                 create: (BuildContext context) => FarmBloc(
                   walletRepository: context.read<WalletRepository>(),
