+++ lib/app/widgets/widgets.dart
@@ -1,0 +1,4 @@
+export 'bottom_navigation_bar_mobile.dart';
+export 'bottom_navigation_bar_web.dart';
+export 'top_navigation_bar_mobile.dart';
+export 'top_navigation_bar_web.dart';
